# projects/phoenix-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Phoenix Platform API
  description: |
    Phoenix Platform API for managing observability cost optimization through
    experiments, pipeline deployments, and metrics analysis.
  version: 2.0.0
  contact:
    name: Phoenix Team
    email: phoenix@example.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://phoenix.example.com/api/v1
    description: Production
security:
  - bearerAuth: []
paths:
...
  /pipelines/deployments:
    get:
      summary: List pipeline deployments
      operationId: listDeployments
      parameters:
        - name: namespace
          in: query
          schema:
            type: string
        - name: pipeline
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of deployments
          content:
            application/json:
              schema:
                type: object
                properties:
                  deployments:
                    type: array
                    items:
                      $ref: '#/components/schemas/PipelineDeployment'
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer

    post:
      summary: Create pipeline deployment
      operationId: createDeployment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeploymentRequest'
      responses:
        '201':
          description: Deployment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineDeployment'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PipelineDeployment:
      type: object
      properties:
        id:
          type: string
        deployment_name:
          type: string
        pipeline_name:
          type: string
        namespace:
          type: string
        variant:
          type: string
        target_nodes:
          type: object
          additionalProperties:
            type: string
        parameters:
          type: object
          additionalProperties: true
        resources:
          $ref: '#/components/schemas/ResourceRequirements'
        status:
          type: string
          enum: [pending, deploying, ready, failed, deleting, degraded, healthy]
        phase:
          type: string
        metrics:
          $ref: '#/components/schemas/DeploymentMetrics'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        created_by:
          type: string
    CreateDeploymentRequest:
      type: object
      required:
        - deployment_name
        - pipeline_name
        - target_nodes
      properties:
        deployment_name:
          type: string
        pipeline_name:
          type: string
        namespace:
          type: string
          default: default
        variant:
          type: string
        target_nodes:
          type: object
          additionalProperties:
            type: string
        parameters:
          type: object
          additionalProperties: true
        resources:
          $ref: '#/components/schemas/ResourceRequirements'
...
    ResourceRequirements:
      type: object
      properties:
        cpu:
          type: string
          example: "100m"
        memory:
          type: string
          example: "128Mi"
        storage:
          type: string
          example: "1Gi"
    DeploymentMetrics:
      type: object
      properties:
        metrics_per_second:
          type: number
        cardinality_reduction:
          type: number
        error_rate:
          type: number
        latency_p99:
          type: number
...